pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-content-starter"
version = "0.1.0"
description = "LLM Accuracy Checker using src/ layout and console scripts"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Tony Morales"}]
dependencies = [
  "beautifulsoup4>=4.12.2",
  "lxml>=5.2.1",
  "markdown-it-py>=3.0.0",
  "sentence-transformers>=2.7.0",
  "faiss-cpu>=1.8.0.post1",
  "numpy>=1.26.4",
  "scikit-learn>=1.4.2",
  "torch>=2.2.0",
  "transformers>=4.41.0",
  "rapidfuzz>=3.9.3",
  "python-dateutil>=2.9.0",
  "streamlit>=1.37.0"
]

[project.urls]
homepage = "https://github.com/tmorales2000/ai-content-starter"

[project.scripts]
accchk-build-index = "accchk.cli:build_index_main"
accchk-check-answer = "accchk.cli:check_answer_main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["accchk*"]
exclude = ["tests*"]
